backend/requirements-vercel.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
google-generativeai==0.3.2
requests==2.31.0
